ml_project/requirements.txt
click==8.1.3
marshmallow_dataclass==8.5.8
gdown==4.4.0
numpy>=1.21.2
pandas==1.1.5
pytest==7.1.2
PyYAML==6.0
scikit_learn==1.0.2
setuptools==58.0.4